[Starts #163683116] refactor code
