+++ app/__init__.py
@@ -2,7 +2,8 @@
 import os
 from flask import Flask
 from instance.config import app_config
-from .v1 import routes
+from .v1.views.base_view import bp
+from .v1.views import offices, parties
 
 
 def create_app(config_name):
@@ -13,6 +14,6 @@ def create_app(config_name):
     app.config.from_pyfile('config.py')
 
     # register blueprints
-    app.register_blueprint(routes.bp)
+    app.register_blueprint(bp)
 
     return app
